Time: 20 ms (40.38%), Space: 6.1 MB (10.94%) - LeetHub

+++ 3-longest-substring-without-repeating-characters/3-longest-substring-without-repeating-characters.c
@@ -6,7 +6,7 @@ int lengthOfLongestSubstring(char * s){
     uint8_t* character = (int*)calloc(128, sizeof(uint8_t));
     for(int i=0; i< len; i++){
         memset(character, 0, 128 * sizeof(uint8_t));
-        int size = 0;
+        uint8_t size = 0;
         for(int j=i; j<len; j++){
             if(character[s[j]]) break;
             character[s[j]] = 1;
